+++ CHANGELOG.md
@@ -18,7 +18,8 @@
   - `video`
     - [#443](https://github.com/wix-incubator/rich-content/pull/443) async url resolving support
     - [#444](https://github.com/wix-incubator/rich-content/pull/444) file upload support
-    - [#449](https://github.com/wix-incubator/rich-content/pull/444) force video mime type for uploads
+    - [#449](https://github.com/wix-incubator/rich-content/pull/449) force video mime type for uploads
+    - [#451](https://github.com/wix-incubator/rich-content/pull/451) uses unique file input id
   ### :bug: Bug Fix
   - `divider`
     - [#438](https://github.com/wix-incubator/rich-content/pull/438) container alignment